Fixed how notes are saved, at the expense of performance.
Now the save button looks up the task when it's clicked rather than looking it up once when it's first loaded. This way, it works when you manually created a task after opening the thread. Unfortunately, this makes the whole thing a bit slower because it has to look the task up each time.