Store data stubs in group store until promise resolves
Currently, if we are in a group's general settings page and a hard
reload (shift-R) is requested, the code always sends the user to the new
groups page. This is because the data store is initially empty, until
the API requests have successfully been resolved, at which point the
group store is finally populated with actual data.

Since LoginPromises now return the groupids that the user belongs to, we
can instead populate the group data store with stub data. This will
prevent the code sending the user to the new groups page.