Adds a new list selector for the all-some-none pattern Also changes how this pattern works. Pattern uses an object like this: `{ all: boolean; some: string[]; none: boolean }` to filter tagged items (usually events).
Previously, pattern worked like this.
* All - Include all items (regardless of tags)
* Some - Include items with at least one tag in this
  list.
* None - Include items with no tags at all.

In particular, "all" was inclusive of both some and none.
That is, all=true implies none=true and that X is in the
some list for any given some.

With this commit, "all" now only implies that we match
for any given tag BUT NOT if there are no tags at all. This
makes it simple to do matching for all tags while still
giving user flexibility to toggle off showing the empty
set.