Refactor Modal and ModalPanel components
Purpose of refactoring is to give us more flexibility on using
Components.Modal with custom headers and footers. Previously,
we just rendered the "footer" inside the body, but this makes
styling with CSS somewhat messy. So now we use the `ModalBase`
component to allow adding custom footers on the same level as
the the body elements.

ModalPanel has been removed in favor of just rendering the
ModalPanelFooter directly since all of the existing uses of
ModalPanel was just to get the footer rendering code.