Simplified layout system
We currently have three separate React root nodes that we can render to --
header, main, and footer, which pre-supposes the DOM looks something like
this:

```
<div id="esper-header">{ ... header content ... }</div>
<div id="esper-main">{ ... main content ... }</div>
<div id="esper-footer">{ ... footer content ...}</div>
```

This isn't ideal because it limits how elements in the main content
can react to contents in the header. In particular, if the header has
release notes, we want the margin-top of the content in `#esper-main`
to increase by the height of the release notes.

So now what we do is target a single React parent node for both the
header and main content (footer is left alone for now). We also
introduce a new `Views.App` wrapper that adds things like the
header and release notes, which can be used like so:

```
Layout.render(<Views.App>
  <Views.MyView {...props} />
</Views.App>)
```

You can always call `Layout.render` without the `<Views.App>` wrapper,
which gives us some flexibility in the future for tweaking page layout
on certain views.