Simplified CSS spacing helpers
Expanded Containers
-------------------

`.esper-full-screen` has been rolled into `esper-expanded`. This
class is used to expand an element to the edges of the containing
element (useful for absolute positioning). The `minus-*` classes
are no longer required. Use adjacent sibling selectors instead.

Sibling selectors are preferred because they reduce the number
of places in React that we have to update to handle sibling
interactions. For example, suppose we have an optional header
followed by a content area. Both are absoutely positioned. The
header should take up some X pixels if present and the content
area takes up the remainder. By using sibling selectors, we can
vary the top margin of the content area based solely on whether
the header is present and not have to worry about applying the
right "minus-header" class to the content area.

Sections
--------

`.esper-section` and `.esper-panel-section` now have vertical
padding and apply negative vertical margins to prevent margin
from stacking when nesting. When in doubt, stick everything in
`.esper-section` and everything will be about 10px apart. Or
stick everything in `.esper-panel-section`s and things will
be about 20px apart with a thin grey line between things.

We previously relied on some other helper elements to add
vertical spacing in places, but in an effort to simply, most of
this spacing logic is now controlled by the `-section` classes.

We also previously had different spacing behavior based on
whether an `.esper-section` was inside a modal or panel. We now
ignore that for the most part. You can use an `.esper-section`
as a direct descendant of a panel body or modal body to reset
the spacing, at which point spacing logic is determined by
the `@esper-base-padding` variable and not the spacing
variables for modals and panels.