Update charts view CSS
We previously used a flex box to expand the chart area to fill the remainder
of the screen height. The problem is that we don't really want the chart to
have a fixed height so much as a min-height. That is, a pie chart should
fill up the remainder of the screen vertically, but a bar chart should keep
extending downward. Using flexbox in the latter case causes the child bar
chart element to overflow outside of the parent, which can lead to some odd
absolute positioning behavior since the parent is shorter than the child.

So instead of flexbox, we should use `calc` instead.