GroupBys, new charting helpers
This is a pretty thorough rewrite of `Charting.tsx`. I would ignore
the diff and go through the file from scratch. The only things that
haven't really changed are the Highcharts helper functions at the
very end.

What we're doing here is defining a bunch of `Types.GroupBy`. Each
`GroupBy` represents an event attribute we're grouping that event
by (like labels, ratings, etc.)

We also replace our previous usage of a Path function with the
`GroupBy` object for things like identifying which filters to apply
and which path to go to.