Bug fixes related to spinner going forever when changing subscriptions.
Now, any promise passed to the `Components.Plans` select callback
that resolves will turn off the spinner. There are places where the
intent is to *not*  turn off the spinner (e.g. when performing a
redirect), but we simply just don't resolve the promises there.

Actions.Subscriptions no longer responsible for redirect. If redirect
required, just invoke it on promise resolution.