Fix simulataneous batch bug
This bug arises when a JsonHttp.batch call is made before a previous
batch call can be resolved.

The batch function relies on a global `batchDfd` deferred object.
The purpose of the global is so that *nested* batch calls can return
a promise that only resolves when the top-level / parent batch call
returns.

However, the global deferred object is reset whenever a new top-level
batch call is made. Because the callback on the first batch function
references the global deferred object, the effect is that completion
of the first batch call will trigger the resolution of the deferred
object that a second ongoing batch call relies upon. This can
result in either (a) an undefined response being passed to a promise
resolution or (b) batch promises never resolving because they're
waiting on a deferred that is no longer referenced correctly.

Solution is simply to have JsonHttp.batch create a locally-scoped
copy of the global variable and have any callbacks within that
function reference only the locally scoped variable.